test(desktop): split import flow specs
