+++ scripts/check-file-sizes.mjs
@@ -69,7 +69,6 @@ const ALLOWLIST = new Set([
   // ── Test files (allow-listed; not a Phase 4 priority) ──────────
   "apps/desktop/src/__tests__/components/timeline/AgentTreeView.test.ts",
   "apps/desktop/src/__tests__/composables/useCachedFetch.test.ts",
-  "apps/desktop/src/__tests__/composables/useImportFlow.test.ts",
   "apps/desktop/src/__tests__/views/analytics-views.test.ts",
 ]);
